backend/requirements.txt
asgiref==3.4.1
astroid==2.8.5
Django==3.2.9
djangorestframework==3.12.4
isort==5.10.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg==3.0.3
psycopg2-binary==2.9.2
pylint==2.11.1
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.13.3
